Refactor chatwidget orchestration into modules (phase 5)
